+++ dependencies/freadoctet.h
@@ -1,0 +1,6 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+void printOctet(char buffer[], int length);
+
+char * readOctet(FILE * file, int nbOctet);
